remove dtype mention

+++ pennylane/pauli/dla/lie_closure.py
@@ -126,7 +126,7 @@ def lie_closure(
             for op in generators
         ]
 
-    vspace = PauliVSpace(generators, tol=tol, dtype=float)
+    vspace = PauliVSpace(generators, tol=tol)
 
     epoch = 0
     old_length = 0  # dummy value
